+++ geapi/src/main/java/io/github/cdsap/geapi/ExperimentReport.kt
@@ -1,6 +1,8 @@
 package io.github.cdsap.geapi
 
+import io.github.cdsap.geapi.domain.GetMeasurements
 import io.github.cdsap.geapi.domain.impl.GetBuildScansWithQueryImpl
+import io.github.cdsap.geapi.domain.impl.GetMeasurementsImpl
 import io.github.cdsap.geapi.domain.impl.GetOutcomeReportImpl
 import io.github.cdsap.geapi.domain.impl.PrintExperimentResultsImpl
 import io.github.cdsap.geapi.domain.model.Filter
@@ -10,12 +12,11 @@ import io.github.cdsap.geapi.view.OutcomeView
 class ExperimentReport(val filter: Filter, val repository: GradleRepositoryImpl) {
 
     suspend fun process() {
-
         val getBuildScans = GetBuildScansWithQueryImpl(repository)
         val getReport = PrintExperimentResultsImpl(repository)
+        val measurements = GetMeasurementsImpl(repository)
         val buildScansFiltered = getBuildScans.get(filter)
-        getReport.print(buildScansFiltered, filter)
-
+        getReport.print(measurements.get(buildScansFiltered, filter))
     }
 }
